добавляем компоненты для списка карточек на странице offer, module5-task1
