+++ src/const/stock-points-count.ts
@@ -1,0 +1,2 @@
+// TODO - удалить после работы с апи
+export const MAX_NEAREST_POINTS_COUNT = 3;
