+++ src/types/app-props.ts
@@ -1,8 +1,9 @@
 import { OfferReview, RentOffer } from './offer.ts';
-import { City } from './map.ts';
+import { City, Point } from './map.ts';
 
 export type AppProps = {
   offers: RentOffer[];
   offerReviews: OfferReview[];
   city: City;
+  points: Point[];
 }
